# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bluebird private aviation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-code:
    post:
      operationId: requestLoginCode
      tags: [auth]
      summary: Request a 6-digit SMS sign-in code for a phone number (demo — code returned in response and logged, no real SMS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCodeRequest"
      responses:
        "200":
          description: Code issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCodeResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Resend cooldown active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: SMS delivery not configured (production — no provider wired; demo delivery is non-production only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-code:
    post:
      operationId: verifyLoginCode
      tags: [auth]
      summary: Verify a phone + 6-digit code, creating the account on first sign-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong, expired, or exhausted code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Revoke the caller's current token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignOutResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SignOutResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RequestCodeRequest:
      type: object
      properties:
        phone:
          type: string
          description: Phone number in any common format; normalized server-side.
      required:
        - phone
    RequestCodeResponse:
      type: object
      properties:
        message:
          type: string
        phone:
          type: string
          description: The normalized phone number the code was issued for.
        demoCode:
          type: string
          description: Demo only — the raw 6-digit code in lieu of a delivered SMS.
        expiresInSeconds:
          type: number
        resendCooldownSeconds:
          type: number
      required:
        - message
        - phone
    VerifyCodeRequest:
      type: object
      properties:
        phone:
          type: string
        code:
          type: string
          description: The 6-digit SMS code.
      required:
        - phone
        - code
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        isNewUser:
          type: boolean
          description: True when this verification created the account.
      required:
        - token
        - user
    UpdateMeRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        membershipTier:
          type: string
          enum: [base, plus, concierge]
        linePassCount:
          type: number
        referralCode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - membershipTier
        - linePassCount
        - referralCode
        - createdAt
...