# lib/api-zod/src/generated/types/requestCodeResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bluebird private aviation API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestCodeResponse {
  message: string;
  /** The normalized phone number the code was issued for. */
  phone: string;
  /** Demo only — the raw 6-digit code in lieu of a delivered SMS. */
  demoCode?: string;
  expiresInSeconds?: number;
  resendCooldownSeconds?: number;
}