# lib/api-zod/src/generated/types/updateMeRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bluebird private aviation API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMeRequest {
  name?: string;
  email?: string;
}