# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bluebird private aviation API
 * OpenAPI spec version: 0.1.0
 */
import type { UserMembershipTier } from './userMembershipTier';

export interface User {
  id: string;
  name: string;
  phone: string;
  email?: string | null;
  membershipTier: UserMembershipTier;
  linePassCount: number;
  referralCode: string;
  createdAt: string;
}